requirements.txt
python-telegram-bot[job-queue]==21.6
requests==2.*
beautifulsoup4==4.*
